requirements.txt
numpy
pandas
pytorch
scikit-learn
allennlp
gensim